fixed qtextedit tooltips and transformation in view_image

+++ qualcoder/view_image.py
@@ -665,7 +665,7 @@ class DialogCodeImage(QtWidgets.QDialog):
         if self.pixmap is None:
             return
         self.scale = (self.ui.horizontalSlider.value() + 1) / 100
-        height = self.scale * self.pixmap.height()
+        height = int(self.scale * self.pixmap.height())
         pixmap = self.pixmap.scaledToHeight(height, QtCore.Qt.TransformationMode.FastTransformation)
         pixmap_item = QtWidgets.QGraphicsPixmapItem(pixmap)
         pixmap_item.setPos(0, 0)
@@ -1663,7 +1663,7 @@ class DialogViewImage(QtWidgets.QDialog):
         if self.pixmap is None:
             return
         scale = (self.ui.horizontalSlider.value() + 1) / 100
-        height = scale * self.pixmap.height()
+        height = int(scale * self.pixmap.height())
         pixmap = self.pixmap.scaledToHeight(height, QtCore.Qt.TransformationMode.FastTransformation)
         pixmap_item = QtWidgets.QGraphicsPixmapItem(pixmap)
         pixmap_item.setPos(0, 0)
